Add Client, MultiFetch, and TagColors
Client contains some helper functions that are not properly associated with
a particular library.

MultiFetch is a helper for loading objects that require multiple requests
to complete, either by continuing to fetch the next batch when it's required,
or by taking care of making all the requests and resolving a promise once
the full set has been retrieved.

TagColors encapsulates the tag color preferences for a library, and matching
those against a passed array of tags, as used to generate colored tag icons
for each item in the web library.