convert Net to use the fetch API
Use whatwg-fetch and node-fetch to polyfill, and fetch-mock to test.
Zotero.ApiResponse now takes an instance of Response as an argument to its
constructor.
Replace use of Deferred module with our own Deferred on top of es6-promise
since we were only using it to hold a ref in a queue and resolve it
outside of the promise body.